--- client/src/Pages/CountryContainer.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-function CountryContainer() {
-
-  return (
-    <h1>Country CountryContainer</h1>
-  )
-}
-
-export default CountryContainer;
