--- client/src/containers/TestMainBox.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Quiz from "../components/Quiz";
-
-// container to test components in development
-const TestMainBox = () => {
-
-
-
-    return (
-    <Quiz/>
-    )
-}
-
-export default TestMainBox;
