+++ client/src/services/PopulateCountries.js
@@ -33,4 +33,4 @@ const PopulateCountries = () => {
 
 }
 
-export default PopulateCountries
+export default PopulateCountries;
